requirements.txt
Flask==0.10.1
requests==2.8.0
newspaper3k==0.2.8
flask_restful
uwsgi